Move web into internal package
