+++ internal/web.go
@@ -1,5 +1,4 @@
-// Package common provides some utility functions for dealing with HTTP requests and responses
-package web
+package internal
 
 import (
 	"bufio"
@@ -9,6 +8,7 @@ import (
 	"io/ioutil"
 	"net/http"
 	"net/textproto"
+	"os"
 	"strings"
 )
 
@@ -55,35 +55,34 @@ func MakeRequest(method string, url string, input io.Reader, headersIncluded boo
 	return http.DefaultTransport.RoundTrip(req)
 }
 
-// WriteRequest writes an http.Request to the specified writer
-func WriteRequest(w io.Writer, req *http.Request, includeMethod bool, includeUrl bool, includeHeaders bool) error {
+// PrintRequest writes an http.Request to stdout
+func PrintRequest(req *http.Request, includeMethod bool, includeUrl bool, includeHeaders bool) error {
 	body, err := ioutil.ReadAll(req.Body)
 	req.Body.Close()
-
 	if err != nil {
 		return err
 	}
 
 	if includeMethod {
-		fmt.Fprintln(w, req.Method)
+		fmt.Println(req.Method)
 	}
 
 	if includeUrl {
-		fmt.Fprintln(w, req.URL)
+		fmt.Println(req.URL)
 	}
 
 	if includeHeaders {
-		req.Header.Write(w)
-		fmt.Fprintln(w)
+		req.Header.Write(os.Stdout)
+		fmt.Println()
 	}
 
-	fmt.Fprintln(w, string(body))
+	fmt.Println(string(body))
 
 	return nil
 }
 
-// WriteResponse writes an http.Response to the specified writer
-func WriteResponse(w io.Writer, resp *http.Response, includeHeaders bool, includeStatus bool) error {
+// PrintResponse writes an http.Response to stdout
+func PrintResponse(resp *http.Response, includeHeaders bool, includeStatus bool) error {
 	body, err := ioutil.ReadAll(resp.Body)
 	resp.Body.Close()
 
@@ -92,15 +91,15 @@ func WriteResponse(w io.Writer, resp *http.Response, includeHeaders bool, includ
 	}
 
 	if includeStatus {
-		fmt.Fprintln(w, resp.Status)
+		fmt.Println(resp.Status)
 	}
 
 	if includeHeaders {
-		resp.Header.Write(w)
-		fmt.Fprintln(w)
+		resp.Header.Write(os.Stdout)
+		fmt.Println()
 	}
 
-	fmt.Fprintln(w, string(body))
+	fmt.Println(string(body))
 
 	return nil
 }
